refactor : separate baseball game number and add test code
