+++ src/main/java/BaseballGameLogic.java
@@ -1,70 +1,7 @@
-import java.util.HashSet;
-import java.util.Set;
-
 public class BaseballGameLogic {
-	private StringBuilder number;
-	private StringBuilder previousNumber;
-	private final int retryLimit = 5;
-	private int retryCount = 0;
-
-	public void makeNumber() {
-		if (!isFirstTry()) {
-			setPreviousNumber();
-			setUnduplicateNumber();
-
-			return;
-		}
-
-		setNumber();
-	}
-
-	private void setUnduplicateNumber() {
-		while (previousNumber.equals(number) && retryCount < retryLimit) {
-			setNumber();
-			retryCount++;
-		}
-
-		retryCount = 0;
-	}
-
-	private boolean isFirstTry() {
-		return previousNumber == null ? true : false;
-	}
-
-	private void setNumber() {
-		int singleNumber;
-		Set<Integer> usedNumbers = new HashSet<>();
-		this.number = new StringBuilder();
-
-		for (int i = 0; i < 3; i++) {
-			singleNumber = getUnduplicateSingleDigit(usedNumbers);
-
-			number.append(singleNumber);
-			usedNumbers.add(singleNumber);
-		}
-	}
-
-	private void setPreviousNumber() {
-		this.previousNumber = this.number;
-	}
-
-	private int getUnduplicateSingleDigit(Set<Integer> usedNumbers) {
-		int candidateNumber = 0;
-
-		while (candidateNumber <= 0 || candidateNumber >= 10 || usedNumbers.contains(candidateNumber)) {
-			candidateNumber = (int) (Math.random() * 10);
-		}
-
-		return candidateNumber;
-	}
-
-	public String getNumber() {
-		return this.number.toString();
-	}
-
-	public String getScore(String input) {
-		int strikeCount = getStrikeCount(input);
-		int ballCount = getBallCount(input);
+	public String getScore(String baseballGameNumber, String guessNumber) {
+		int strikeCount = getStrikeCount(baseballGameNumber, guessNumber);
+		int ballCount = getBallCount(baseballGameNumber, guessNumber);
 
 		if (strikeCount == 0 && ballCount == 0) {
 			return "낫싱";
@@ -87,7 +24,7 @@ public class BaseballGameLogic {
 		return result.toString();
 	}
 
-	private int getBallCount(String input) {
+	private int getBallCount(String number, String input) {
 		int ballCount = 0;
 
 		for (int i = 0; i < 3; i++) {
@@ -98,7 +35,7 @@ public class BaseballGameLogic {
 		return ballCount;
 	}
 
-	private int getStrikeCount(String input) {
+	private int getStrikeCount(String number, String input) {
 		int strikeCount = 0;
 
 		for (int i = 0; i < 3; i++) {
